Add XML files to parse and prep dummy test file

+++ src/FoldingRenderer/Models/Panel.cs
@@ -7,7 +7,7 @@ namespace FoldingRenderer.Models
     public Guid Id { get; set; }
     public string Name { get; set; }
     public PanelRotation Rotation { get; set; }
-    public int HingeOffset { get; set; }
+    public double HingeOffset { get; set; }
     public PanelDimensions Dimensions { get; set; }
     public int AttachedToSide { get; set; }
     public PanelCrease Crease { get; set; }
